for customzing a GPU only build
This will build the container with both CPU and GPU inferencing.  it refers to Nvidia as the target GPU infrastructure but you can modify this file to support a different GPU as ollama adds more GPUs to its support.

This yml has the same effect as the standard docker-compose.yml which builds for CPU and GPU (nvidia) support.